convert ExpandingTextField to typescript
